backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0

pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.0


python-docx==1.1.0
PyPDF2==3.0.1
google-generativeai
openpyxl
